Add test_calcAirDensity unit test. (#298)
Add test_calcAirDensity unit test. This highlighted that 1.0e+0_DP is preferable to 1.0d0 syntax, so that is changed in calcAirDensity(). Other instances of this should now be changed in future, and the perl script looking for 'single-precision numbers' should be dropepd, or replaced by one that hunts for 'non-_DP' numbers instead. Slightly loosen relative tolerance for tests.