Catch missing RO2 species error when prepping
Adds an exception where an RO2 species is encountered in the RO2 sum of
the mechanism file but isn't encountered in the mechanism species list.

We shouldn't get to a point where there is an RO2 species in
mechanism.ro2 that isn't in the full species list in the mechanism, so
the user needs to correct their mechanism file before running the model!